requirements.txt
fastapi
uvicorn
python-dotenv
openai
tiktoken
# For fuzzy topic matching
thefuzz
# For file uploads
python-multipart

